requirements.txt
gradio
transformers
torch
diffusers
spacy
openpyxl
huggingface_hub
sentence-transformers
pandas
tabulate
llama-index
chromadb
langchain
accelerate
pillow
python-dotenv
ffmpeg
SpeechRecognition
pydub
ffmpeg-python
imageio
imageio-ffmpeg
gtts
fastapi
uvicorn
python-multipart